New chart api + new overview stock fragment menu

+++ app/src/main/java/com/example/stocks/models/remote/StockDailyPrice.kt
@@ -8,7 +8,7 @@ data class StockDailyPrice(
     @SerializedName("open") var open: Double? = null,
     @SerializedName("high") var high: Double? = null,
     @SerializedName("low") var low: Double? = null,
-    @SerializedName("close") var close: Double? = null,
+    @SerializedName("close") var close: Float? = null,
     @SerializedName("adjClose") var adjClose: Double? = null,
     @SerializedName("volume") var volume: Int? = null,
     @SerializedName("unadjustedVolume") var unadjustedVolume : Int? = null,
